Satisfy clang-tidy and the 32 bit legs
The bucket types held their fingerprints and counters in C arrays, which the tidy
configuration refuses; std::array now, read through data() for the vector load.
The allocator-extended move constructor of group_storage moves member by member,
which the same annotation as the table's marks. And the churn test indexed a
vector with a 64 bit random value, a narrowing only a 32 bit size_t can see.

+++ test/unit/group_bucket.cpp
@@ -183,7 +183,7 @@ TEST_CASE("group_index_churn_at_full_load") {
     REQUIRE(map.bucket_count() == buckets); // reserve was enough, no growth
     for (size_t round = 0; round < 4; ++round) {
         for (size_t i = 0; i < num_elements; ++i) {
-            auto const slot = rng() % live.size();
+            auto const slot = static_cast<size_t>(rng() % live.size());
             REQUIRE(map.erase(live[slot]) == 1U);
             REQUIRE(map.try_emplace(next, i).second);
             live[slot] = next;
